Refactoring, correciones y terminados los estados de Mario, menos la estrella

+++ src/entities/updateables/MovableEntity.java
@@ -2,10 +2,10 @@ package entities.updateables;
 
 public interface MovableEntity {
     public void switchDirection();
-    public int getSpeedY();
-    public void setSpeedY(int y);
-    public int getSpeedX();
-    public void setSpeedX(int x);
+    public float getSpeedY();
+    public void setSpeedY(float y);
+    public float getSpeedX();
+    public void setSpeedX(float x);
     public void applyGravity();
 }
